+++ HISTORY.md
@@ -1,5 +1,11 @@
 History of features added to LiChess Tools in time (reversed order)
 
+26 Jan
+  - PGN Editor keep found, undo, redo
+
+25 Jan
+  - Nicer flairs theme
+
 22 Jan
   - Eval rows in Explorer
   - FEN search in PGN Editor